// src/app/shared/Models/catalogos/campoticketcaseta.model.ts
export class CampoTicketCaseta {
    id: number = 0;
    nombre: string = "";
    descripcion: string = "";
    tipoDatoId: number = 71;
    casetaId : number = 0;
    fechaAlta: string = "0001-01-01";
    bActivo: boolean = true;
    nombreTipoDato: string = "";
    bEditable: boolean = false;
    bNuevoCampo: boolean = false;


    constructor(){}
}